# docs/swagger.yaml
basePath: /
definitions:
...
  dto.FilmResponse:
    properties:
      based_on_scenario:
        type: boolean
      created_at:
        type: string
      director:
        type: string
      film_type:
        type: string
      id:
        type: string
      release_year:
        type: integer
      title:
        type: string
      video_path:
        type: string
    type: object
...
  handler.JSONResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API for Hemra Şirow literary and biography website
  title: Literary Backend API
  version: "1.0"
paths:
...
  /api/films:
    get:
      description: List all films
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.FilmResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: List films
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Create a new film. Supports JSON or multipart/form-data with optional
        video upload.
      parameters:
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Film type (film or animation)
        in: formData
        name: film_type
        required: true
        type: string
      - description: Release year
        in: formData
        name: release_year
        type: integer
      - description: Director
        in: formData
        name: director
        type: string
      - description: Based on scenario
        in: formData
        name: based_on_scenario
        type: boolean
      - description: Video file
        in: formData
        name: file
        type: file
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FilmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Create film
  /api/films/{id}:
    delete:
      description: Delete an existing film by id
      parameters:
      - description: Film ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Delete film
    get:
      description: Get a single film by id
      parameters:
      - description: Film ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FilmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Get film
    put:
      consumes:
      - application/json
      - multipart/form-data
      description: Update an existing film. Supports JSON or multipart/form-data with
        optional video upload.
      parameters:
      - description: Film ID
        in: path
        name: id
        required: true
        type: string
      - description: Title
        in: formData
        name: title
        type: string
      - description: Film type (film or animation)
        in: formData
        name: film_type
        type: string
      - description: Release year
        in: formData
        name: release_year
        type: integer
      - description: Director
        in: formData
        name: director
        type: string
      - description: Based on scenario
        in: formData
        name: based_on_scenario
        type: boolean
      - description: Video file
        in: formData
        name: file
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FilmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Update film
...
schemes:
- http
- https
swagger: "2.0"